+++ internal/app/storage/storage_test.go
@@ -1,11 +1,12 @@
 package storage
 
 import (
-	"github.com/stretchr/testify/assert"
-	"go.uber.org/zap"
 	"log"
 	"os"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"go.uber.org/zap"
 
 	"github.com/sergalkin/go-url-shortener.git/internal/app/config"
 )
@@ -46,8 +47,37 @@ func TestNewStorage(t *testing.T) {
 		})
 	}
 
-	err := os.Remove("tmp")
-	if err != nil {
-		log.Fatalln(err)
+	if _, fErr := os.Stat("tmp"); fErr == nil {
+		err := os.Remove("tmp")
+		if err != nil {
+			log.Fatalln(err)
+		}
+	}
+}
+
+func TestNewStorageCanCreateDbStorage(t *testing.T) {
+	tests := []struct {
+		name string
+		do   func()
+	}{
+		{
+			name: "DB will be created if DSN is provided",
+			do: func() {
+				config.NewConfig(
+					config.WithDatabaseConnection("postgres://root:root@localhost:5432/postgres?sslmode=disable"),
+				)
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			tt.do()
+			s, err := NewStorage(&zap.Logger{})
+			if err == nil {
+				assert.NotNil(t, s)
+				assert.IsType(t, &db{}, s)
+				assert.NotEmpty(t, s)
+			}
+		})
 	}
 }
